WIP abstract product
